Separate Validators into their own top-level src folder
