+++ src/controllers/Admin-Web-App/administrative/variable-controller.js
@@ -8,7 +8,7 @@ const {
     MakeNewVariableValidation, 
     DeleteVariableValidation, 
     UpdateStatusValidation, 
-} = require('../../../middlewares/form-validation/variable-validators')
+} = require('../../../validations/variable-validators')
 const { 
     findVariableParentQuery,  
     findVariableQuery,
